core: guard size and payload calculations
Bound allocation, payload, plane, packet, frame, header, and metadata size calculations before narrowing or writing.

Squashes original commits 536-650 after 4c03ab28.

+++ common/set.c
@@ -120,9 +120,9 @@ int x264_cqm_init( x264_t *h )
         }\
         else\
         {\
-            CHECKED_MALLOC( h->  quant##w##_mf[i], (QP_MAX_SPEC+1)*size*sizeof(udctcoef) );\
-            CHECKED_MALLOC( h->dequant##w##_mf[i],  6*size*sizeof(int) );\
-            CHECKED_MALLOC( h->unquant##w##_mf[i], (QP_MAX_SPEC+1)*size*sizeof(int) );\
+            CHECKED_MALLOC( h->  quant##w##_mf[i], (int64_t)(QP_MAX_SPEC+1)*size*sizeof(udctcoef) );\
+            CHECKED_MALLOC( h->dequant##w##_mf[i],  (int64_t)6*size*sizeof(int) );\
+            CHECKED_MALLOC( h->unquant##w##_mf[i], (int64_t)(QP_MAX_SPEC+1)*size*sizeof(int) );\
         }\
         for( j = 0; j < i; j++ )\
             if( deadzone[j] == deadzone[i] &&\
@@ -135,8 +135,8 @@ int x264_cqm_init( x264_t *h )
         }\
         else\
         {\
-            CHECKED_MALLOC( h->quant##w##_bias[i], (QP_MAX_SPEC+1)*size*sizeof(udctcoef) );\
-            CHECKED_MALLOC( h->quant##w##_bias0[i], (QP_MAX_SPEC+1)*size*sizeof(udctcoef) );\
+            CHECKED_MALLOC( h->quant##w##_bias[i], (int64_t)(QP_MAX_SPEC+1)*size*sizeof(udctcoef) );\
+            CHECKED_MALLOC( h->quant##w##_bias0[i], (int64_t)(QP_MAX_SPEC+1)*size*sizeof(udctcoef) );\
         }\
     }
 
@@ -223,7 +223,7 @@ int x264_cqm_init( x264_t *h )
 
     /* Emergency mode denoising. */
     x264_emms();
-    CHECKED_MALLOC( h->nr_offset_emergency, sizeof(*h->nr_offset_emergency)*(QP_MAX-QP_MAX_SPEC) );
+    CHECKED_MALLOC( h->nr_offset_emergency, (int64_t)sizeof(*h->nr_offset_emergency)*(QP_MAX-QP_MAX_SPEC) );
     for( int q = 0; q < QP_MAX - QP_MAX_SPEC; q++ )
         for( int cat = 0; cat < 3 + CHROMA444; cat++ )
         {
